+++ src/main/java/roomescape/service/reservation/ReservationDeleteService.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.service.reservation;
 
 import org.springframework.stereotype.Service;
 import roomescape.repository.ReservationRepository;
